noroute: disable systemd-timesyncd due to problems and use ntp instead
systemd-timesyncd gave an error while starting up after rebuiling that
/var/lib is not a directory. It also tried to use a path inside
/nix/store as the state directory. This resulted in the clock getting
stuck on the 1970s. Which in turn resulted that no SSL certificates were
valid yet, which blocked using 'nixos-rebuild switch'  as well.

For this to be fixed you had to first set the date manually (in order to
SSL to work)

    # date -s '2019-11-15 12:00'

Then nixos-rebuild could be used again to fetch from cache.nixos.org.