hardware/wifi/dlink-dwa-172.nix: fix configuration, create override
Most of the overlays have been compacted into the
nix/overlays/kernel-packages.nix and into the nix/overlays/packages.nix
file instead of using a seperate file for each package.

Additionally the build for the rtl8821au module has been activated on
aarch64-linux as well in order to be able to try to build it on the
raspberry pi.